Remove comment link from meta section of articles
Who really cares how many comments there are on a post and that a
reader can quickly jump to the comments. If someone wants to see the
comments they will scroll to the bottom of the page.

Sure, maybe there are some people who are enticed by seeing number of
comments on the main page but oh well. Better to not have the visual
clutter for something that probably most people don't care about.